backend/requirements.txt
Django==4.2.23
djangorestframework==3.14.0
django-cors-headers==4.3.1
openai==1.12.0
python-dotenv==1.0.1
gunicorn==23.0.0
google-generativeai==0.3.1
requests==2.31.0
psycopg2-binary==2.9.9 
# Added by auto-fixer
firebase-admin>=6.0.0
groq>=0.31.0
python-decouple>=3.8
whitenoise>=6.5.0
